+++ crates/flipperzero/src/furi/thread.rs
@@ -19,12 +19,11 @@ use alloc::{
 use flipperzero_sys as sys;
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 const MIN_STACK_SIZE: usize = 1024;
 
 /// Thread factory, which can be used in order to configure the properties of a new thread.
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 pub struct Builder {
     /// Guaranteed to be UTF-8.
     name: Option<CString>,
@@ -33,7 +32,6 @@ pub struct Builder {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl Builder {
     /// Generates the base configuration for spawning a thread, from which configuration
     /// methods can be chained.
@@ -144,7 +142,7 @@ impl Builder {
 /// This call will create a thread using default parameters of [`Builder`]. If you want to
 /// specify the stack size or the name of the thread, use that API instead.
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 pub fn spawn<F>(f: F) -> JoinHandle
 where
     F: FnOnce() -> i32,
@@ -155,7 +153,7 @@ where
 
 /// Gets a handle to the thread that invokes it.
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 pub fn current() -> Thread {
     use alloc::borrow::ToOwned;
 
@@ -199,12 +197,12 @@ pub fn sleep(duration: core::time::Duration) {
 
 /// A unique identifier for a running thread.
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 pub struct ThreadId(sys::FuriThreadId);
 
 /// A handle to a thread.
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 pub struct Thread {
     /// Guaranteed to be UTF-8.
     name: Option<CString>,
@@ -212,7 +210,6 @@ pub struct Thread {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl Thread {
     fn new(
         name: Option<CString>,
@@ -268,7 +265,6 @@ impl Thread {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl fmt::Debug for Thread {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("Thread")
@@ -278,7 +274,6 @@ impl fmt::Debug for Thread {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl ufmt::uDebug for Thread {
     fn fmt<W>(&self, f: &mut ufmt::Formatter<'_, W>) -> Result<(), W::Error>
     where
@@ -291,13 +286,12 @@ impl ufmt::uDebug for Thread {
 
 /// An owned permission to join on a thread (block on its termination).
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 pub struct JoinHandle {
     context: Option<Arc<Thread>>,
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl Drop for JoinHandle {
     fn drop(&mut self) {
         let context = self
@@ -315,7 +309,6 @@ impl Drop for JoinHandle {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl JoinHandle {
     /// Extracts a handle to the underlying thread.
     pub fn thread(&self) -> &Thread {
@@ -341,7 +334,6 @@ impl JoinHandle {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl fmt::Debug for JoinHandle {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("JoinHandle").finish_non_exhaustive()
@@ -349,7 +341,6 @@ impl fmt::Debug for JoinHandle {
 }
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "alloc")))]
 impl ufmt::uDebug for JoinHandle {
     fn fmt<W>(&self, f: &mut ufmt::Formatter<'_, W>) -> Result<(), W::Error>
     where