+++ crates/flipperzero/src/lib.rs
@@ -1,40 +1,49 @@
 //! High-level bindings for the Flipper Zero.
+//!
+//! # Features
 #![cfg_attr(feature = "document-features", doc = document_features::document_features!())]
+//!
+
 #![no_std]
 #![cfg_attr(test, no_main)]
 #![cfg_attr(feature = "unstable_intrinsics", feature(int_roundings))]
 #![cfg_attr(feature = "unstable_lints", feature(must_not_suspend))]
-#![cfg_attr(feature = "unstable_docs", feature(doc_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
+#![deny(rustdoc::broken_intra_doc_links)]
 
-#[cfg(feature = "alloc")]
+#[cfg(any(feature = "alloc", docsrs))]
+#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))]
 extern crate alloc;
 
 #[cfg(feature = "service-dialogs")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-dialogs")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-dialogs")))]
 pub mod dialogs;
 #[cfg(feature = "service-dolphin")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-dolphin")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-dolphin")))]
 pub mod dolphin;
 pub mod furi;
 #[cfg(feature = "service-gui")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-gui")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-gui")))]
 pub mod gui;
 #[cfg(feature = "service-input")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-input")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-input")))]
 pub mod input;
 pub(crate) mod internals;
 pub mod io;
 pub mod kernel;
 pub mod macros;
 #[cfg(feature = "service-notification")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-notification")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-notification")))]
 pub mod notification;
 #[cfg(feature = "service-storage")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-storage")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-storage")))]
 pub mod storage;
 #[cfg(feature = "service-toolbox")]
-#[cfg_attr(feature = "unstable_docs", doc(cfg(feature = "service-toolbox")))]
+#[cfg_attr(docsrs, doc(cfg(feature = "service-toolbox")))]
 pub mod toolbox;
+#[cfg(feature = "xbm")]
+#[cfg_attr(docsrs, doc(cfg(feature = "xbm")))]
+pub mod xbm;
 
 #[doc(hidden)]
 pub mod __macro_support {
@@ -76,6 +85,6 @@ flipperzero_test::tests_runner!(
         crate::toolbox::crc32::tests,
         crate::toolbox::md5::tests,
         crate::toolbox::sha256::tests,
-        crate::gui::xbm::tests,
+        crate::xbm::tests,
     ]
 );