+++ crates/sys/src/lib.rs
@@ -1,6 +1,7 @@
 //! Low-level bindings for the Flipper Zero.
 
 #![no_std]
+#![deny(rustdoc::broken_intra_doc_links)]
 
 // Features that identify thumbv7em-none-eabihf.
 // Until target_abi is stable, this also permits thumbv7em-none-eabi.
@@ -17,9 +18,15 @@ core::compile_error!("This crate requires `--target thumbv7em-none-eabihf`");
 pub mod furi;
 mod inlines;
 
-#[allow(non_upper_case_globals)]
-#[allow(non_camel_case_types)]
-#[allow(non_snake_case)]
+#[allow(
+    non_upper_case_globals,
+    non_camel_case_types,
+    non_snake_case,
+    clippy::missing_safety_doc,
+    clippy::transmute_int_to_bool,
+    clippy::useless_transmute,
+    rustdoc::broken_intra_doc_links
+)]
 mod bindings;
 
 /// Create a static C string.