+++ tools/src/bin/storage.rs
@@ -1,8 +1,8 @@
 //! Storage CLI.
 //!
-//! See: [storage.py][1] script.
+//! See: [`storage.py`][1] script.
 //!
-//! [1]: (https://github.com/flipperdevices/flipperzero-firmware/blob/dev/scripts/storage.py)
+//! [1]: https://github.com/flipperdevices/flipperzero-firmware/blob/dev/scripts/storage.py
 
 use std::path::PathBuf;
 use std::process;
@@ -85,7 +85,7 @@ fn main() {
 
     let port_info =
         serial::find_flipperzero(cli.port.as_deref()).expect("unable to find Flipper Zero");
-    let port = serialport::new(&port_info.port_name, serial::BAUD_115200)
+    let port = serialport::new(port_info.port_name, serial::BAUD_115200)
         .timeout(Duration::from_secs(30))
         .open()
         .expect("unable to open serial port");
@@ -109,7 +109,7 @@ fn main() {
         Commands::Receive {
             flipper_path,
             local_path,
-        } => store.receive_file(flipper_path, &local_path),
+        } => store.receive_file(flipper_path, local_path),
         Commands::Send {
             local_path,
             flipper_path,